% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MutHeatmap}
\alias{MutHeatmap}
\title{Mutation heatmap with multiple mutation types}
\usage{
MutHeatmap(vr, pal = c("#F2F2F2", colorRampPalette(c("blue", "white",
  "red"))(5)[c(1, 2)], "#F2F2F2", colorRampPalette(c("blue", "white",
  "red"))(5)[c(4, 5)], brewer.pal(n = 8, name = "Accent")[c(1, 4, 6, 8, 2,
  3, 5, 7)], "#E31A1C", "#6A3D9A"), type = c("DEL", "LOSS", "NEUTRAL",
  "GAIN", "AMPL", "nonsynonymous SNV", "synonymous SNV", "intronic",
  "stopgain", "nonframeshift deletion", "splicing", "frameshift deletion",
  "UTR3", "frameshift insertion", "UTR5"), order_gene = T,
  order_patient = T, hist_plot = T, legend_dist = 0.4,
  col_text_cex = 1, row_text_cex = 1, sub_gene = NULL,
  heatmap_mar = c(5, 17, 1, 2), heatmap_oma = c(0.2, 0.2, 0.2, 0.2),
  heatmap_mex = 0.5, legend_mar = c(1, 0, 4, 1), xlab_adj = 1,
  order_omit = c("NEUTRAL"), annotation_col = NULL,
  annotation_colors = NULL, heatmap_height = 3, heatmap_width = 3,
  anno_height = NULL)
}
\arguments{
\item{vr}{The data frame containing the variance data, there are three columns,
the column names should be: Gene, Type, Patient}

\item{pal}{Character vectors, Need to be customized according to the number of
mutation types in the data frame, need more than the mutation type one color, it
is used as the background color, and the background color is placed in the
first place}

\item{type}{Character vectors. The corresponding mutation type, type have to equal
to or more than all types appearing in the data; by default, the four types of copy
number plus neutral plus annovar mutations. Default: "DEL","LOSS","NEUTRAL","GAIN",
"AMPL","nonsynonymous SNV","synonymous SNV","intronic","stopgain","nonframeshift
deletion", "splicing", "frameshift deletion","UTR3","frameshift insertion","UTR5"
mutations annovar; the length of type is less  1 than pal}

\item{order_gene}{A logical scalar. Default T, sorted according to the number of
patients gene mutations}

\item{order_patient}{A logical scalar. Default T, according to the patient The number of mutated
genes is sorted}

\item{hist_plot}{A logical scalar.Default T, plus the corresponding histogram above and to the right}

\item{legend_dist}{Numeric. Default 0.4, adjust the distance between the legends, generally need to
adjust}

\item{col_text_cex}{Numeric. Adjust the size of the patient name, default. 1}

\item{row_text_cex}{Numeric. Gene name adjust the size of the default. 1}

\item{sub_gene}{Numeric. Character vectors. Drawing only selected portions of genes
and the gene is present in the data needed, the default NULL}

\item{heatmap_mar}{Numeric. Mar parameter, adjusting the length of the left and right edges
of the front and rear heat FIG default C ( 5,17,1,2)}

\item{legend_mar}{Numeric. Legend Mar parameter, adjust the position of the legend, default
c (1,0,4,1)}

\item{xlab_adj}{Numeric. Patient name and adjustment FIG distance between}

\item{order_omit}{Character vectors. The type of variation that is ignored when sorting, these mutation
types will also be filtered in the histogram, the default c ("NEUTRAL"), If there is
 no mutation type "NEUTRAL", you can also keep the default parameters}

\item{heatmap_height}{Numeric. The height of the heat map. For the case of a large number of
genes, based on the length of the pdf, you can set the heatmap to cover the height of
 the canvas. Grid}

\item{heatmap_width}{Numeric. The width of the heat map. For a very large number of patients,
you can set the length of the heat map}

\item{Heatmap_oma}{Numeric. Oma parameter, adjust the outer edge length of the heat map before
and after, the default c (0.2, 0.2, 0.2, 0.2)}

\item{mex}{Numeric. Adjust the heat map's mex parameter, used to describe the coordinates of
the drawing edge, the default 0.5}

\item{Annotation_col}{Data.frame. Information annotating the patient,  the line name
 is the patient, the column name is different pathological information, the data must be
  a factor, rownames should be patients. The example is as follows (in the code example I briefly SmokingInfo is Smoking,
   OldInfo is Old)
       SmokingInfo   OldInfo
P1     Smoking   Old
P2  NonSmoking Unold
P3  NonSmoking Unold
P4  NonSmoking   Old
P5     Smoking Unold
P6  NonSmoking Unold}

\item{Annotation_colors}{Character vectors list. Match different color to pathological information, examples are
as follows, need to match the column name and factor in annotation_col
$SmokingInfo
Smoking NonSmoking
"#FDB462"  "#80B1D3"
$OldInfo
Old     Unold
"#FB8072" "#8DD3C7"}

\item{Anno_height}{Sets the height of the annotation, which is automatically adjusted by default.}
}
\description{
Plot a heat map with multiple mutation types within a gene
}
\details{
The gene heat map that we usually see is a grid of genes, a type of mutation,
but in fact the same gene often has multiple mutation types in the same patient,
so the traditional heat map drawing tool can't satisfy us. For the problem, this
 function uses image to draw the preliminary heat map, and then uses points to
add the second mutation in the form of a square. The third mutation is added in
turn. At the same time, the square position is slightly moved and accompanied by
a slight reduction in size to achieve a better display effect, and up to four
mutations can be represented on one heat map grid.

Before run MutHeatmap, one should known:

1. If you want to draw a heat map with a histogram, because the image size is too large, use the pdf
function and give it a large enough width and length;

2. The default is the type of mutation annotated by annovar;

3. Because there are too many factors affecting the alignment of the heat map and the histogram during
drawing, it is difficult to adjust the corresponding mar, mex, and oma parameters to achieve better
results. Therefore, it is recommended to quickly draw a rough estimate and then use inkscape or
adobe for layout alignment.

4. If the point of the mutation type in the heat map is too small, the width and length of the pdf file
should be reduced.
}
\examples{
#without hist plot
pdf("heatmap_cnv_mut.pdf", height=12, width = 12)
MutHeatmap(vr, heatmap_mar = c(17,17,1,2),hist_plot = F, legend_dist=0.1, xlab_adj = 1.2, order_patient = T, order_gene = T)
dev.off()

#with hist plot
pdf("heatmap_hist_cnv_mut.pdf", height=12, width = 12)
MutHeatmap(vr, heatmap_mar = c(17,7,1,2),hist_plot = T, legend_dist=0.3, xlab_adj = 1.2, order_patient = T, order_gene = T)
dev.off()

#only a few gene
pdf("Assoc_CN1.pdf", height=2,width = 14)
# have to adjust legend by manual
MutHeatmap(vr, heatmap_mar = c(7,17,1,2), sub_gene = c("CDKN2A", "GNAQ", "NOTCH1", "RB1", "SMAD4", "ABL1"),hist_plot = F,legend_dist=0.2, xlab_adj = 0.9)
dev.off()

#with annotation and hist
annotation_col <- data.frame(Smoking = factor(sample(c("Smoking", "NonSmoking"), length(unique(vr$Patient)), replace = T)), Old=factor(sample(c("Old", "Unold"), length(unique(vr$Patient)), replace = T)))
rownames(annotation_col) <- unique(vr$Patient)
annotation_colors <- list(Smoking =c(Smoking = "#FDB462", NonSmoking = "#80B1D3"), Old=c(Old = "#FB8072", Unold = "#8DD3C7"))
pdf("heatmap_hist_cnv_mut.pdf", height=15, width = 12)
MutHeatmap(vr, heatmap_mar = c(15,10,1,2), legend_mar = c(1,0,1,1), hist_plot = T, legend_dist=0.2, xlab_adj = 1.2, order_patient = T, order_gene = T, annotation_col=annotation_col, annotation_colors=annotation_colors)
dev.off()

#with annotation and without hist
pdf("heatmap_cnv_mut.pdf", height=12, width = 12)
MutHeatmap(vr, heatmap_mar = c(17,10,1,2),hist_plot = F, legend_dist=0.1, xlab_adj = 1.2, order_patient = T, order_gene = T, annotation_col=annotation_col, annotation_colors=annotation_colors)
dev.off()

#Start with the first step(there is not test data in MutHeatmap package)
# read variant data
dt <- read.csv("~/Downloads/selected_genetic_2.csv", sep = ";")
setDT(dt)
setnames(dt, c("Patient", "Gene", "Type")) # change column name
# read clinical data
ClinicalInfo <- read.csv("~/Downloads/selected_tma_clinic_2.csv", sep = ";")
annotation_col <- ClinicalInfo[,-1]
rownames(annotation_col) <- ClinicalInfo[,1]
annotation_colors <- list(Age = c("<50" = "#FDB462", ">50" = "#80B1D3"), Smoking=c("yes" = "#FB8072", "no" = "#8DD3C7"),PDL1_status = c("pos" = "#3288BD", "neg" = "#5E4FA2") )
# plot heatmap
pdf("example_4.pdf", height=15, width = 12)
MutHeatmap(dt, pal = c("#F2F2F2",brewer.pal(8, "Paired")) ,type =
            c("missense", "nonsense", "frameshiftDeletion", "frameshiftInsertion", "nonframeshiftDeletion", "Amplification", "Deletion", "Fusion"),order_gene = T, order_patient = T, hist_plot =
            T, legend_dist = 0.4, col_text_cex = 1, row_text_cex = 1, sub_gene= NULL,heatmap_mar = c(20,17,1,2), heatmap_oma=c(0.2,0.2,0.2,0.2),heatmap_mex=0.5, legend_mar = c(1,0,4,1),xlab_adj=0.2,
          order_omit=c("NEUTRAL"), annotation_col = annotation_col, annotation_colors = annotation_colors, heatmap_height = 5,
          heatmap_width = 6, anno_height = 0.5)
dev.off()

}
