% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr.R
\docType{data}
\name{vr}
\alias{vr}
\title{Variants of lung cancer}
\format{A data table with 2512 rows and 3 variables}
\usage{
data(vr)

* `Gene`: Gene
* `Type`: Mutation type
* `Patient`: Patient id
}
\description{
A dataset containing the patients, copy number variants and mutations
vr. The variables are as follows:
}
\keyword{datasets}
